# API/Swagger.yaml
openapi: 3.0.1
info:
  title: AwesomeSoft API
  description: >-
    AwesomeSoft API documentation
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: simonsondrup@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'https://localhost:5000/api'
  - url: 'http://localhost:5000/api'
...
paths:
...
  /participant:
    put:
      tags:
        - participant
      summary: Update an existing participant
      operationId: updateParticipant
      requestBody:
        description: Participant update object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Participant'
          application/xml:
            schema:
              $ref: '#/components/schemas/Participant'
        required: true
      responses:
        '400':
          description: Invalid ID supplied
          content: {}
        '404':
          description: Participant not found
          content: {}
        '405':
          description: Validation exception
          content: {}
      x-codegen-request-body-name: body
    post:
      tags:
        - participant
      summary: Add a new participant
      description: Create a new participant
      operationId: addParticipant
      requestBody:
        description: Participant object that needs to be created
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Participant'
          application/xml:
            schema:
              $ref: '#/components/schemas/Participant'
        required: true
      responses:
        '405':
          description: Invalid input
          content: {}
      x-codegen-request-body-name: body
  '/participant/{participantId}':
    get:
      tags:
        - participant
      summary: Find participant by ID
      description: Returns a single participant
      operationId: getParticipantById
      parameters:
        - name: participantId
          in: path
          description: ID of participant to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Participant'
            application/json:
              schema:
                $ref: '#/components/schemas/Participant'
        '400':
          description: Invalid ID supplied
          content: {}
        '404':
          description: Participant not found
          content: {}
      security:
        - api_key: []
    delete:
      tags:
        - participant
      summary: Deletes a participant
      operationId: deleteParticipant
      parameters:
        - name: participantId
          in: path
          description: participant id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid ID supplied
          content: {}
        '404':
          description: Participant not found
          content: {}
...
components:
  schemas:
...
    Participant:
      required:
        - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: Simon
      xml:
        name: Participant
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
          scopes:
            'write:pets': modify pets in your account
            'read:pets': read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header